+++ examples/not-found.ts
@@ -1,0 +1,17 @@
+
+import { Fluent } from '@moebius/fluent';
+
+
+void (async () => {
+
+  const fluent = new Fluent();
+
+  await fluent.addTranslation({
+    locales: 'en-GB',
+    source: '',
+  });
+
+  // {welcome}
+  console.debug(fluent.translate('en', 'welcome'));
+
+})();